Fix Cyclic Import, add dispatch
