+++ ingestion/src/metadata/ingestion/source/database/column_type_parser.py
@@ -150,7 +150,7 @@ class ColumnTypeParser:
         "MEDIUMTEXT": "MEDIUMTEXT",
         "MONEY": "NUMBER",
         "NCHAR": "CHAR",
-        "NTEXT": "TEXT",
+        "NTEXT": "NTEXT",
         "NULL": "NULL",
         "NUMBER": "NUMBER",
         "NUMERIC": "NUMERIC",
@@ -285,23 +285,20 @@ class ColumnTypeParser:
 
         _COLUMN_TYPE_MAPPING[BIT] = "BINARY"
         _SOURCE_TYPE_TO_OM_TYPE["BIT"] = "BINARY"
+        _SOURCE_TYPE_TO_OM_TYPE["IMAGE"] = "IMAGE"
     except ImportError:
         pass
 
     @staticmethod
     def get_column_type(column_type: Any) -> str:
-        column_type_result = ColumnTypeParser.get_column_type_mapping(column_type)
-        if column_type_result:
-            return column_type_result
-        column_type_result = ColumnTypeParser.get_source_type_mapping(column_type)
-        if column_type_result:
-            return column_type_result
-        column_type_result = ColumnTypeParser.get_source_type_containes_brackets(
-            column_type
-        )
-        if column_type_result:
-            return column_type_result
-
+        for func in [
+            ColumnTypeParser.get_column_type_mapping,
+            ColumnTypeParser.get_source_type_mapping,
+            ColumnTypeParser.get_source_type_containes_brackets,
+        ]:
+            column_type_result = func(column_type)
+            if column_type_result:
+                return column_type_result
         return ColumnTypeParser._SOURCE_TYPE_TO_OM_TYPE.get("UNKNOWN")
 
     @staticmethod
