+++ ingestion/src/metadata/profiler/metrics/static/distinct_count.py
@@ -18,6 +18,7 @@ Distinct Count Metric definition
 from sqlalchemy import column, distinct, func
 
 from metadata.profiler.metrics.core import StaticMetric, _label
+from metadata.profiler.orm.functions.count import CountFn
 from metadata.utils.logger import profiler_logger
 
 logger = profiler_logger()
@@ -40,7 +41,7 @@ class DistinctCount(StaticMetric):
 
     @_label
     def fn(self):
-        return func.count(distinct(column(self.col.name)))
+        return func.count(distinct(CountFn(column(self.col.name, self.col.type))))
 
     def df_fn(self, dfs=None):
         from collections import Counter  # pylint: disable=import-outside-toplevel
