+++ ingestion/src/metadata/profiler/orm/converter/base.py
@@ -13,82 +13,20 @@
 Converter logic to transform an OpenMetadata Table Entity
 to an SQLAlchemy ORM class.
 """
-
-from abc import abstractmethod
 from typing import Optional, cast
 
-import sqlalchemy
 from sqlalchemy import MetaData
 from sqlalchemy.orm import DeclarativeMeta, declarative_base
 
 from metadata.generated.schema.entity.data.database import Database, databaseService
 from metadata.generated.schema.entity.data.databaseSchema import DatabaseSchema
-from metadata.generated.schema.entity.data.table import Column, DataType, Table
+from metadata.generated.schema.entity.data.table import Table
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
-from metadata.ingestion.source import sqa_types
-from metadata.profiler.orm.registry import CustomTypes
 
 Base = declarative_base()
 
-_TYPE_MAP = {
-    DataType.NUMBER: sqlalchemy.NUMERIC,
-    DataType.TINYINT: sqlalchemy.SMALLINT,
-    DataType.SMALLINT: sqlalchemy.SMALLINT,
-    DataType.INT: sqlalchemy.INT,
-    DataType.BIGINT: sqlalchemy.BIGINT,
-    DataType.BYTEINT: sqlalchemy.SMALLINT,
-    DataType.BYTES: CustomTypes.BYTES.value,
-    DataType.FLOAT: sqlalchemy.FLOAT,
-    DataType.DOUBLE: sqlalchemy.DECIMAL,
-    DataType.DECIMAL: sqlalchemy.DECIMAL,
-    DataType.NUMERIC: sqlalchemy.NUMERIC,
-    DataType.TIMESTAMP: CustomTypes.TIMESTAMP.value,
-    DataType.TIME: sqlalchemy.TIME,
-    DataType.DATE: sqlalchemy.DATE,
-    DataType.DATETIME: sqlalchemy.DATETIME,
-    DataType.INTERVAL: sqlalchemy.Interval,
-    DataType.STRING: sqlalchemy.String,
-    DataType.MEDIUMTEXT: sqlalchemy.TEXT,
-    DataType.TEXT: sqlalchemy.TEXT,
-    DataType.CHAR: sqlalchemy.CHAR,
-    DataType.VARCHAR: sqlalchemy.VARCHAR,
-    DataType.BOOLEAN: sqlalchemy.BOOLEAN,
-    DataType.BINARY: sqlalchemy.LargeBinary,
-    DataType.VARBINARY: sqlalchemy.VARBINARY,
-    DataType.ARRAY: CustomTypes.ARRAY.value,
-    DataType.BLOB: CustomTypes.BYTES.value,
-    DataType.LONGBLOB: sqlalchemy.LargeBinary,
-    DataType.MEDIUMBLOB: sqlalchemy.LargeBinary,
-    DataType.MAP: sqa_types.SQAMap,
-    DataType.STRUCT: sqa_types.SQAStruct,
-    DataType.UNION: sqa_types.SQAUnion,
-    DataType.SET: sqa_types.SQASet,
-    DataType.GEOGRAPHY: sqa_types.SQASGeography,
-    DataType.ENUM: sqlalchemy.Enum,
-    DataType.JSON: sqlalchemy.JSON,
-    DataType.UUID: CustomTypes.UUID.value,
-    DataType.BYTEA: CustomTypes.BYTEA.value,
-}
-
 
 SQA_RESERVED_ATTRIBUTES = ["metadata"]
-
-
-class BaseMapTypes:
-    """
-    Base Class for mapping types
-    """
-
-    def map_types(self, col: Column, table_service_type):
-        """returns an ORM type"""
-
-        if col.arrayDataType:
-            return _TYPE_MAP.get(col.dataType)(item_type=col.arrayDataType)
-        return self.return_custom_type(col, table_service_type)
-
-    @abstractmethod
-    def return_custom_type(self, col: Column, table_service_type):
-        return _TYPE_MAP.get(col.dataType)
 
 
 def check_snowflake_case_sensitive(table_service_type, table_or_col) -> Optional[bool]:
@@ -124,31 +62,6 @@ def check_if_should_quote_column_name(table_service_type) -> Optional[bool]:
     return None
 
 
-def build_orm_col(idx: int, col: Column, table_service_type) -> sqlalchemy.Column:
-    """
-    Cook the ORM column from our metadata instance
-    information.
-
-    The first parsed column will be used arbitrarily
-    as the PK, as SQLAlchemy forces us to specify
-    at least one PK.
-
-    As this is only used for INSERT/UPDATE/DELETE,
-    there is no impact for our read-only purposes.
-    """
-
-    return sqlalchemy.Column(
-        name=str(col.name.__root__),
-        type_=BaseMapTypes().map_types(col, table_service_type),
-        primary_key=not bool(idx),  # The first col seen is used as PK
-        quote=check_if_should_quote_column_name(table_service_type)
-        or check_snowflake_case_sensitive(table_service_type, col.name.__root__),
-        key=str(
-            col.name.__root__
-        ).lower(),  # Add lowercase column name as key for snowflake case sensitive columns
-    )
-
-
 def ometa_to_sqa_orm(
     table: Table, metadata: OpenMetadata, sqa_metadata_obj: Optional[MetaData] = None
 ) -> DeclarativeMeta:
@@ -161,6 +74,8 @@ def ometa_to_sqa_orm(
     `type` and passing SQLAlchemy `Base` class
     as the bases tuple for inheritance.
     """
+    # pylint: disable=import-outside-toplevel,cyclic-import
+    from metadata.profiler.orm.converter.dispatch_converter import build_orm_col
 
     table.serviceType = cast(
         databaseService.DatabaseServiceType, table.serviceType
