+++ ingestion/src/metadata/profiler/orm/converter/snowflake/converter.py
@@ -17,10 +17,10 @@ to an SQLAlchemy ORM class.
 
 from metadata.generated.schema.entity.data.database import databaseService
 from metadata.generated.schema.entity.data.table import Column, DataType
-from metadata.profiler.orm.converter.base import BaseMapTypes
+from metadata.profiler.orm.converter.common import CommonMapTypes
 
 
-class SnowflakeMapTypes(BaseMapTypes):
+class SnowflakeMapTypes(CommonMapTypes):
     def return_custom_type(self, col: Column, table_service_type):
         if (
             table_service_type == databaseService.DatabaseServiceType.Snowflake
